+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[b5e1e44](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/b5e1e4413aa91f1b309e141433cacb4cfdf71941)...[444824d](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/444824dbf2a01dd955fc9cfb645a36fcb15c9d71)
+[b5e1e44](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/b5e1e4413aa91f1b309e141433cacb4cfdf71941)...[ef1bcc2](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/ef1bcc2e2e862f626b748b28bd66567bf45934bb)
 
 ### Bug Fixes
 
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Enhance replace_tag function to return success status and improve error handling ([`05b647d`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/05b647df89da182194cad8ad172e4d37d75d1554))
 - Add --clobber option to gh release create command ([`7da8704`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/7da8704465ea0dfd747c1a55ff274026702644a9))
 - Add --clobber option to gh release upload command ([`3436971`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/343697183d8e14ad63fe2413a93001d59f39992f))
+- Mount model crate migrations dir (#296) ([`ef1bcc2`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/ef1bcc2e2e862f626b748b28bd66567bf45934bb))
 
 ### Documentation
 
@@ -46,6 +47,8 @@ All notable changes to this project will be documented in this file.
 - Update CHANGELOG.md and checksums [skip ci] ([`814be69`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/814be6988cb3c8b6241f69dc9c9cfa6a201e957c))
 - Update CHANGELOG.md and checksums [skip ci] ([`1e08efd`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/1e08efd27fc26f1c5d75ee3e501b5c0fef5d2742))
 - Update CHANGELOG.md and checksums [skip ci] ([`907eb33`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/907eb33f2cabcd745e6037973fb73d6c3c3d1ddc))
+- Update CHANGELOG.md and checksums [skip ci] ([`f2cbb89`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/f2cbb8911f63211c69af7547062ccf4588c4ad02))
+- Update CHANGELOG.md and checksums [skip ci] ([`e65fa39`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/e65fa39edc9e8e7fd7e00d54a1c4ffd6d20420e4))
 
 ### Features
 
@@ -55,6 +58,7 @@ All notable changes to this project will be documented in this file.
 - Replace CSV ledger with PostgreSQL backend (#292) ([`0083133`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/0083133e934e759d0a25c53a420b62b04c137a49))
 - CRL storage to PostgreSQL (#294) ([`7e90546`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/7e90546d7212e8813589c60d81f7e6a3bd7a20c9))
 - Use unwrap infaillable for random number generation and pgnotify() for notification channel ([`24094fe`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/24094fefe60aafcd10152615fdb9f407540ab6da))
+- Spool to PostgreSQL (#296) ([`ee6f1b9`](https://github.com/ADORSYS-GIS/wazuh-cert-oauth2/commit/ee6f1b9b79788591a12e36da244a69ef7d593444))
 
 ### Miscellaneous Tasks